bluetooth connection working using the service
i am using a service to run the accelerometer in the background and
update the data to the main ui thread, it is  than passed to the app
client and over the network.  now i must convert the data to an int[]
and use the readings to control the cursor.